refactor express server, run as child process
