+++ src/main/expressServer.ts
@@ -1,49 +1,9 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import express from 'express';
-import fs from 'fs';
+import { ChildProcess, spawn } from 'child_process';
+import path from 'path';
 
-const app = express();
+export default function startExpressServer(): ChildProcess {
+  const pathToServer = path.join(process.cwd(), 'src', 'expressServer');
 
-const streamVideo: (req: any, res: any) => void = (req, res) => {
-  const encodeFilePath = req.params.name;
-  const sourcePath = Buffer.from(encodeFilePath, 'base64').toString('utf-8');
-  const stat = fs.statSync(sourcePath);
-  const fileSize = stat.size;
-  const { range } = req.headers;
-
-  if (range) {
-    const parts = range.replace(/bytes=/, '').split('-');
-    const start = parseInt(parts[0], 10);
-    const end = parts[1] ? parseInt(parts[1], 10) : fileSize - 1;
-
-    if (start >= fileSize) {
-      res
-        .status(416)
-        .send(`Requested range not satisfiable\n${start} >= ${fileSize}`);
-      return;
-    }
-
-    const chunksize = end - start + 1;
-    const file = fs.createReadStream(sourcePath, { start, end });
-    const head = {
-      'Content-Range': `bytes ${start}-${end}/${fileSize}`,
-      'Accept-Ranges': 'bytes',
-      'Content-Length': chunksize,
-      'Content-Type': 'video/mp4',
-    };
-
-    res.writeHead(206, head);
-    file.pipe(res);
-  } else {
-    const head = {
-      'Content-Length': fileSize,
-      'Content-Type': 'video/mp4',
-    };
-    res.writeHead(200, head);
-    fs.createReadStream(sourcePath).pipe(res);
-  }
-};
-
-app.get('/video/:name', streamVideo);
-
-app.listen(process.env.EXPRESS_PORT);
+  return spawn('node', ['server.js'], { cwd: pathToServer });
+}
